requirements.txt
langchain
ipykernel
langchain_groq
langchain
python-dotenv
langchain-google-genai 
langchain_community
langchain-text-splitters
google-generativeai
chromadb
langchain_huggingface
sentence-transformers
langchain_core
streamlit